fio: enable overlap checking with offload submission
Enable use of the serialize_overlap option with
io_submit_mode=offload. Threads with both of these options enabled will
check all offloaded in-flight IOs for overlap. If overlap is found,
submission of the new IO will be held back until the overlapping IO has
completed. Threads must be used when fio is run in this mode.

This fixes "fio serialize across jobs"
https://www.spinics.net/lists/fio/msg07133.html