Revert "blktrace: support for non-512b sector sizes"
This reverts commit 26e616d054f238bf6b7ada351e17b72fdfbbb8e0.

An issue appeared when replaying blktraces using fio on non-512b sector
size devices where fio would report the following when replaying a trace
back on the same device.

  fio: io_u error on file /dev/sdb: No space left on device:

This turns out to be due to a match of how blktrace exports trace
information and how fio interprets it. The kernel considers all IO to be
relative to 512b sectors, so the sector offset is always in these units.
However, when replaying fio was adapting to the hardware sector. This
caused offset calculation issues resulting in the above error.

Reverting this keeps fio more in line with expectations, but will force
users who want to have fio run workloads be cognizant of aligning IO
properly when running on specific media to get expected performance.

Reported-by: Vernon Miller <vwm@fb.com>
Signed-off-by: Dennis Zhou <dennis@kernel.org>
Signed-off-by: Jens Axboe <axboe@kernel.dk>