verify: add requested block information to failure trace
If verify_interval is not equal to the block size, FIO would verify the
data one segment by one segment according to the verify_interval. The
offset and length in the verify failure trace are the offset and length
of the corrupted segment, not the offset and block size of the IO.

On block storage system, we need the offset and length of the IO to
investigate data corruption issue. For me, I usually search the offset
of the IO in the storage system trace to figure out what happened in the
system.

Fixes: https://github.com/axboe/fio/issues/709
Signed-off-by: Jens Axboe <axboe@kernel.dk>