filesetup: don't call create_work_dirs() for ioengine with FIO_DISKLESSIO
Don't create work dirs if ioengine has flag FIO_DISKLESSIO. So that
filename can contains '/'. It is useful when ioengine is based on
network and filename is used as url.

Also call create_work_dirs() just when setting up a file. So that
ioengine is loaded and ioengine flags are available.

Signed-off-by: Kyle Zhang <kyle@smartx.com>