nbd: Update for libnbd 0.9.8
As the libnbd API isn't permanently stable until we reach the 1.0
release (expected soon), some code changes are needed to cope with API
changes between 0.9.6 and 0.9.8.  In this case we made changes to
completion handlers after feedback from reviewers.  This fix for fio
incorporates all the changes needed and bumps the minimum version to
libnbd >= 0.9.8.

Signed-off-by: Richard W.M. Jones <rjones@redhat.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>