engines/io_uring: fix crash with registerfiles=1
If used with a raw bdev, we're crashing in attempting to open a
registered file before we have actually registered them.

If we're called before files are registered, just open the file
normally. This is done to query sizes etc, and we'll get the file
closed after that anyway. The job open/close will use the right
registered fd.

Signed-off-by: Jens Axboe <axboe@kernel.dk>