t/run-fio-tests: a script to automate running fio tests
fio actually has a substantial test infrastructure but lacks a
convenient way to run existing tests in an automated fashion. This
script provides the means to run almost all of fio's existing tests with
a single command. A full run finishes in less than 10 minutes.

This script can be used to catch bugs before patches are submitted. It
can also be used to test new fio builds as patches are committed to the
repository.